pyproject.toml
[tool.poetry]
name = "2fa-microservice"
version = "0.1.0"
description = ""
authors = ["MJedr <jedrych.marcysia@gmail.com>"]
packages = [
    { include = "*", from = "app" },
]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.104.1"
starlette = "<0.28.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
exceptiongroup = "^1.1.3"
structlog = "^23.2.0"
httpx = "^0.25.0"
pytest-httpx = "^0.26.0"
uvicorn = "^0.23.2"
pytest-vcr = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
mock = "^5.1.0"
pytest-asyncio = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
